requirements.txt
Flask==3.0.0
Flask-JWT-Extended==4.6.0
Flask-Limiter==3.5.0
Flask-CORS==4.0.0
psycopg2-binary==2.9.9
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
ollama==0.1.6
python-dotenv==1.0.0
Werkzeug==3.0.1
